finishing MainActivity/network requesting and create RestaurantActivity
i finish the MainActivity (1°restaurant choosen informations by means of Retrofit request for going to RestaurantActivity / 2° alarm notifications) , i create the RestaurantActivity (accessible from the correspondant button on the NavigationDrawerMenu on MainActivity)  who displays the restaurant informations choosen by the user (also i create the WebViewActivity for displaying the internet site of the restaurant) , with the Adapter and the ViewHolder for the RecyclerView displayed in the RestaurantActivity. I also finishing all i need for network requesting (i finish the creation of all the pojo class that i needed, i create the Retrofit Object who is reponsible for executing network request). also i create some utils for retrieve date and hours. now i will begin to create the SettingActivity (second button on the NavigationDrawer on the MainActivity), some other utils for the alarm. Next i will create the fragments for the BottomNavigationView