Implementation of Firebase storage
now you can store image on Firebase storage , get their URL and save it on Firebase Firestore.
with that you can really send a message with an image in the chat (and the image is store on firebase).
before you could only recover an image from your phone and see an preview in the chat.